Refactor sunrise into separate component
